libgo: revert nelems/freeindex/freeIndexForScan to uintptr (gc 1.22 allocheaders uint16 corrupts gccgo fork heap → GC zombie crash)

+++ libgo/go/runtime/export_test.go
@@ -1416,7 +1416,7 @@ func FreeMSpan(s *MSpan) {
 
 func MSpanCountAlloc(ms *MSpan, bits []byte) int {
 	s := (*mspan)(ms)
-	s.nelems = uint16(len(bits) * 8)
+	s.nelems = uintptr(len(bits) * 8)
 	s.gcmarkBits = (*gcBits)(unsafe.Pointer(&bits[0]))
 	result := s.countAlloc()
 	s.gcmarkBits = nil
